Move the FakeLLM in tests/unit_tests from there to llms, and refactor some other implementations of FakeLLM
