+++ tests/unit_tests/chains/test_api.py
@@ -8,7 +8,7 @@ from langchain import LLMChain
 from langchain.chains.api.base import APIChain
 from langchain.chains.api.prompt import API_RESPONSE_PROMPT, API_URL_PROMPT
 from langchain.requests import RequestsWrapper
-from tests.unit_tests.llms.fake_llm import FakeLLM
+from langchain.llms.fake_llm import FakeLLM
 
 
 class FakeRequestsChain(RequestsWrapper):
