+++ tests/unit_tests/chains/test_hyde.py
@@ -8,6 +8,7 @@ from langchain.chains.hyde.base import HypotheticalDocumentEmbedder
 from langchain.chains.hyde.prompts import PROMPT_MAP
 from langchain.embeddings.base import Embeddings
 from langchain.llms.base import BaseLLM
+from langchain.llms.fake_llm import FakeLLM
 from langchain.schema import Generation, LLMResult
 
 
@@ -21,22 +22,6 @@ class FakeEmbeddings(Embeddings):
     def embed_query(self, text: str) -> List[float]:
         """Return random floats."""
         return list(np.random.uniform(0, 1, 10))
-
-
-class FakeLLM(BaseLLM, BaseModel):
-    """Fake LLM wrapper for testing purposes."""
-
-    n: int = 1
-
-    def _generate(
-        self, prompts: List[str], stop: Optional[List[str]] = None
-    ) -> LLMResult:
-        return LLMResult(generations=[[Generation(text="foo") for _ in range(self.n)]])
-
-    @property
-    def _llm_type(self) -> str:
-        """Return type of llm."""
-        return "fake"
 
 
 def test_hyde_from_llm() -> None:
