+++ tests/unit_tests/chains/test_llm_math.py
@@ -4,7 +4,7 @@ import pytest
 
 from langchain.chains.llm_math.base import LLMMathChain
 from langchain.chains.llm_math.prompt import _PROMPT_TEMPLATE
-from tests.unit_tests.llms.fake_llm import FakeLLM
+from langchain.llms.fake_llm import FakeLLM
 
 
 @pytest.fixture
